formatting

+++ include/medium/validate_binary_search_tree.h
@@ -6,13 +6,14 @@
 #include "../treenode.h"
 
 namespace Leet::Medium {
-    // Given the root of a binary tree, determine if it is a valid binary search tree
-    // (BST).
+    // Given the root of a binary tree, determine if it is a valid binary search
+    // tree (BST).
 
     // A valid BST is defined as follows:
-    // The left subtree of a node contains only nodes with keys less than the node's
-    // key. The right subtree of a node contains only nodes with keys greater than the
-    // node's key. Both the left and right subtrees must also be binary search trees.
+    // The left subtree of a node contains only nodes with keys less than the
+    // node's key. The right subtree of a node contains only nodes with keys
+    // greater than the node's key. Both the left and right subtrees must also
+    // be binary search trees.
 
     // Example 1:
     // Input: root = [2,1,3]
@@ -43,7 +44,8 @@ namespace Leet::Medium {
             return true;
         }
 
-        void isValidBST_memory_helper(Leet::TreeNode *node, std::vector<int> &arr)
+        void isValidBST_memory_helper(Leet::TreeNode *node,
+                                      std::vector<int> &arr)
         {
             if (!node)
                 return;
@@ -60,7 +62,8 @@ namespace Leet::Medium {
         }
 
         // think of this as a valid "range"
-        bool isValidBST_helper(Leet::TreeNode *node, long min, long max)
+        bool isValidBST_helper(Leet::TreeNode *node, long long min,
+                               long long max)
         {
             if (!node)
                 return true;
